Update integration test to use valid HTML5 (#1780)
This change fixes a test in our suite broken by
https://github.com/rails/rails/pull/48523.

Before the upstream change, invalid the invalid HTML5 in this test
case was parsed without issue. The upstream change switched the
Nokogiri parsing mode to HTML5, leading to parsing issues with
our test case.

I've converted all of the DOM nodes in this case into divs,
preserving the intention of the test case but making it valid HTML5.