Restore API documentation example for with_request_url's `host` param
It looks like this may have been inadvertently deleted in
231d4bb72ce5df334fca6f3be4ed33e887bd2cc8 by regenerating the API docs.
That happened because this example was manually added directly in
`api.md` in 5decd07842c48cbad82527daefa3fe9c65a4226a, rather than in the
corresponding YARD comment. To fix this, we've added the example to the
YARD comment and regenerated the API docs.