Fix slot names that start with call
When a component is registered that starts with `call` the compiler
detects it as a template method which is not correct. This changes the
compiler to look for `call(_|$)` instead of `call` to avoid this
issue.

This means that slots can't (and couldn't) start with `call_` since we
still rely on the `call_` naming convention to generate the
template methods. To make this more dev friendly, this adds a check
that raises an error if a slot name starts with `call_`.

A more long-term fix would be to use some kind of template method
container instead of relying on the `call_` naming convention.

Fixes https://github.com/ViewComponent/view_component/issues/1825