Use component path for generating RSpec files
Generally, the path to an RSpec file should match the app path, but
substituting `app` for `spec`. If a user changes the generation path for
their components, then the generator should use an appropriate path that
matches that. In the example where the user changes the generation path
from `app/components` to `app/views/components`, the RSpec component
spec should be created in `spec/views/components`.

The logic chosen here checks that the component is still being generated
into the `app` dir, otherwise it will just fallback on the usual
destination of `spec/components`.

Since this is a breaking change, it is managed by the
`use_component_path_for_rspec_tests` config option.

NB: Since this includes the `AbstractGenerator`, we can remove the
`file_name` method as it's provided by the module.