Test against turbo-rails v2
hotwired/turbo-rails#294 was closed, which prompted me to check that the issue had been fixed. It's been closed as stale, but this puts us on the latest `turbo-rails` for testing this incompatibility.

Since we're no longer testing against a version of Rails that's incompatible with `turbo-rails`, I've made it a development dependency.