use Struct instead of OpenStruct
I'm guessing that we're seeing failures referencing OpenStruct
here now that Template is in its own class instead of in Compiler,
which includes concurrent-ruby.

As it turns out, we don't even need OpenStruct! We know what
attributes we need at the point of instantiation, so we can just
use Struct, which uses less memory anyways <3