Underscore internal methods that are public
We have plenty of internal methods that are not meant to be
used externally, often tagged @private. In a recent office hours,
@camertron, @blakewilliams and I noted that we should better
indicate when an internal method should not be dependended on.
I've done so here by adding a `__vc_` prefix as is convention
in other parts of the codebase already.

As a separate exercise, we should probably re-evaluate our public
API surface area and see if we can make any public methods private.