Wrap entire compile step with mutex
Currently we only wrap the redefinition of templates with a mutex. This
has largely been fine, but recent changes have highlighted that being
overly specific with the mutex still leads to race conditions that are
complex to debug or test for.

This change wraps the entire compile step with a mutex to help avoid
introducing any race conditions during the compile step.

I tested this change against the last race condition bug and validated
that the demo app I setup locally still runs as expected.