Delegate config to parent modules (no-op) (#2231)
* Delegate config to parent modules

Adds a mechanism for inheriting config from a parent module. There's definitely more that needs to be done to iron this out, including:
- if the module does not already respond to `config`, provide a concern that sets it up using `ActiveSupport::Configurable`
- if the module _does_ already respond to `config` and doesn't have `view_component` set on it, set it up.
- if the module's config does have `view_component` set up, leave it alone?

I also need to consider whether the current way of having things refer back to the Rails app config is appropriate.

Since `Base` is generally used when referring to "global" config options, this probably doesn't do anything right now, but settings like
`test_controller` and `generate` should opt into it.

* Provide ViewComponent::Configurable module

This can be included on a module to provide module-local configuration for ViewComponent. It avoids overwriting existing configuration that may already exist.

* Adjust allocations

* Update changelog