Add #current_template accessor and Template#path (#2283)
Prior to the compiler changes in v4, we set @current_template
to be the current instance of the component. Now, we set it to the
current ViewComponent::Template. In the GitHub monolith, we
reference @current_template when doing certain diagnostic
operations. I added an accessor to make this less messy to use
(and test) and exposed #path on template, which now allows for
such diagnostic operations to reference the specific component
template being rendered.