[v4] BREAKING: Move config.view_component_path to config.generate.path (#2295)
* Move config.view_component_path to config.generate.path

In auditing config.view_component_path for the coming configuration
changes we're making, I realized that our usage of the option was
riddled with bugs.

The issues all had to do with the underlying assumption that there
can only be _one_ view_component_path, when in fact components
can exist in any number of locations.

For example, we used it to strip part of the component file path
for generating translation keys. This of course only worked if the
component was in the directory set in view_component_path.

We also used it to populate the Rails statistics for ViewComponents.
I opted to remove the stats functionality for now as we'd need a
reliable way of counting components outside of a provided directory
path.

* add test case for translations for component in module

* document support for global translations

* Fix final line endings

---------

Co-authored-by: GitHub Actions Bot <41898282+github-actions[bot]@users.noreply.github.com>