Add an example of a feature and system spec
These specs attempts to demonstrate an issue we found porting an old Rails
application to view components with RSpec.

We wanted to render a component in isolation with a feature spec so that
we can test the JavaScript behaviour (via Stimulus in our case).

We felt like this was supported with `with_rendered_component_path` as
shown in the Minitest example:

https://viewcomponent.org/guide/testing.html#component-system-tests

However we could not `visit` the rendered view (in the tmp directory) or
call anything useful as `visit` and other action methods are not
available.

The visit turned out not to be an issue as the
`with_rendered_component_path` appears to load the rendered component:

https://github.com/ViewComponent/view_component/blob/a531d9d604cadce6180b1bd1173f2de631517656/lib/view_component/system_test_helpers.rb#L21

But to test any interactivity provided by JavaScript we would need other actions
like `click_on` which are not available.

The issue appears to stem form `page` being set to a
`Capybara::Node::Simple` which intentionally does not include actions:

https://rubydoc.info/gems/capybara/Capybara/Node/Simple

`page` is set by `ViewComponent::TestHelpers` here:

https://github.com/ViewComponent/view_component/blob/a531d9d604cadce6180b1bd1173f2de631517656/lib/view_component/test_helpers.rb#L11

Not overriding `page` or setting it back to the current
`Capybara::Session` as we do here:

https://github.com/ViewComponent/view_component/blob/main/lib/view_component/system_test_case.rb

is likely to be an answer.

This work does not aim to resolve this issue, merely demonstrate it for
further investigation.