Add a proof of concept for feature and system specs
We have learnt that in Minitest, system specs override `page` back to
the expected `Capybara::Session` so we can do the same in each of our
concept specs to prove this is indeed causing the issue.

The only thing left is to work out how to set the page method back ONLY
for feature and system specs in RSpec.