Potential fix: RSpec system and feature specs
We can use a new module as a way to override the `page` method and set
it back to `Capybara::Session` to allow both RSpec feature and system
specs to work along with `with_rendered_component_path`,

The module includes all the existing `SystemTestHelpers` behaviour and
adds the `page` method in the same way that `SystemTestCase` does.

The proposal would be that for RSpec use the `SystemSpecHelpers` and
for Minitest, use the existing `SystemTestHelpers`.

This is a little clunky and error prone, and I suspect there is a better
way, but this can open up that conversation.