remove Who Uses ViewComponent section
We originally added this section to the docs to provide social
proof of who was adopting ViewComponent. With the framework now
well adopted by community, I think it's time to retire this part
of the docs. It has lost its utility and contributions to the list
muddy up the changelog. I also have a hunch that it's not a great
idea to be linking off to company sites like this for SEO purposes,
as we don't want to imply endorsement of any given corporate entity.